pyproject.toml
[tool.poetry]
name = "hackathon-career-track-711"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
Django = "^4.2.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
djoser = "^2.2.0"
drf-spectacular = "^0.26.5"
gunicorn = "^21.2.0"
pillow = "^10.1.0"
django-cors-headers = "^4.3.0"
django-filter = "^23.3"
pandas = "^2.1.2"
xlsxwriter = "^3.1.9"

[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
ruff = "^0.1.1"
isort = "^5.12.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py311']
exclude = 'migrations'
line_length = 79

[tool.isort]
line_length = 79
skip = ['migrations', 'settings']
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true

[tool.ruff]
target-version = "py311"
exclude = [
    "*migrations/",
    "settings.py",
]
line-length = 79